+++ benchmarks/networks/__init__.py
@@ -1,2 +1,0 @@
-from benchmarks.networks.network import Node, Network
-from benchmarks.networks import creator, tester, modifier, drawer, configuration