+++ simulation/networks/__init__.py
@@ -1,0 +1,2 @@
+from simulation.networks.network import Node, Network
+from simulation.networks import creator, tester, modifier, drawer, configuration