+++ simulation/tests/test_search.py
@@ -1,7 +1,7 @@
 import logging
 import unittest
 
-from benchmarks import networks, utils
+from simulation import networks, utils
 
 log = logging.getLogger(__name__)
 
@@ -11,7 +11,7 @@ class TestCyclicSearch(unittest.TestCase):
         results = networks.configuration.Configuration(
             num_nodes=5,
             connect_type=networks.configuration.ConnectType.CYCLICAL,
-            num_connects=1).run(f"benchmarks_output/tests/"
+            num_connects=1).run(f"simulation_output/tests/"
                                 f"test_cyclic_{utils.get_formatted_time()}")
 
         self.assertEqual(results["percentage_success"], 1)
@@ -22,7 +22,7 @@ class TestRootedSearch(unittest.TestCase):
         results = networks.configuration.Configuration(
             num_nodes=5,
             connect_type=networks.configuration.ConnectType.ROOTED,
-            num_connects=1).run(f"benchmarks_output/tests/"
+            num_connects=1).run(f"simulation_output/tests/"
                                 f"test_rooted_{utils.get_formatted_time()}")
 
         self.assertEqual(results["percentage_success"], 1)